move LanguagePickerSelect to be with the other inputs, because it well, is
