+++ localization/react-intl/src/app/components/cds/inputs/LanguagePickerSelect.json
@@ -1,0 +1,20 @@
+[
+  {
+    "id": "LanguagePickerSelect.optionLabel",
+    "defaultMessage": "{languageName} ({languageCode})"
+  },
+  {
+    "id": "LanguagePickerSelect.unknownLanguage",
+    "defaultMessage": "Unknown language"
+  },
+  {
+    "id": "LanguagePickerSelect.selectLanguage",
+    "description": "Change language label",
+    "defaultMessage": "Select language"
+  },
+  {
+    "id": "LanguagePickerSelect.selectLanguageLabel",
+    "description": "Label for input to select language of fact-check",
+    "defaultMessage": "Language"
+  }
+]
